requirements.txt
jax
jaxlib
dynamaxsys
cbfax
qpax
numpy
matplotlib
equinox
PyQt5
pandas
tqdm
pillow